+++ cases/cbti/index.js
@@ -1,0 +1,5 @@
+export const meta = { id: 'cbti', name: 'CBTI', desc: '猫咪人格测试', author: 'Lovstudio', authorUrl: 'https://github.com/lovstudio' };
+export { dimensionMeta, dimensionOrder, DIM_EXPLANATIONS } from './dimensions';
+export { questions, specialQuestions } from './questions';
+export { CATNIP_TRIGGER_QUESTION_ID as DRUNK_TRIGGER_QUESTION_ID } from './questions';
+export { TYPE_LIBRARY, TYPE_IMAGES, NORMAL_TYPES } from './types';