+++ cases/registry.js
@@ -1,0 +1,4 @@
+import * as sbti from './sbti';
+import * as cbti from './cbti';
+
+export const CASES = [sbti, cbti];