+++ cases/sbti/index.js
@@ -1,0 +1,4 @@
+export const meta = { id: 'sbti', name: 'SBTI', desc: '沙比人格测试', author: '蛆肉儿串儿', authorUrl: 'https://space.bilibili.com/417038183' };
+export { dimensionMeta, dimensionOrder, DIM_EXPLANATIONS } from './dimensions';
+export { questions, specialQuestions, DRUNK_TRIGGER_QUESTION_ID } from './questions';
+export { TYPE_LIBRARY, TYPE_IMAGES, NORMAL_TYPES } from './types';