+++ vite.config.js
@@ -1,7 +1,11 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import { lovinspPlugin } from 'lovinsp'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    lovinspPlugin({ bundler: 'vite' }),
+    react(),
+  ],
 })